Refactor elibc atexit into enclave core
- Remove elibc atexit.c

- Refactor internal/atexit.h into elibc/stdlib.h and enclave/core/sgx/atexit.h

- Refactor libc to use elibc/bits/atexit.h wrapper over oe_atexit
