+++ enclave/core/sgx/atexit.h
