+++ libc/atexit.c
@@ -1,20 +1,18 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include <openenclave/internal/atexit.h>
-#include <stdlib.h>
+#include <openenclave/elibc/stdlib.h>
 
-/*
-**==============================================================================
-**
-** atexit()
-**
-**     Enclave implementation of the libc atexit function.
-**
-**==============================================================================
-*/
-
-int atexit(void (*function)(void))
-{
-    return oe_atexit(function);
-}
+#if !defined(OE_NEED_STDC_NAMES)
+#define OE_NEED_STDC_NAMES
+#define __UNDEF_OE_NEED_STDC_NAMES
+#endif
+#if defined(OE_INLINE)
+#undef OE_INLINE
+#define OE_INLINE
+#endif
+#include <openenclave/elibc/bits/atexit.h>
+#if defined(__UNDEF_OE_NEED_STDC_NAMES)
+#undef OE_NEED_STDC_NAMES
+#undef __UNDEF_OE_NEED_STDC_NAMES
+#endif
